refactor(config): absorb preset defaults into core
