+++ .agents/execution/handoff.json
@@ -2,26 +2,28 @@
   "schemaVersion": 3,
   "handoff": {
     "repository": "DigitalHerencia/LoadedVibes",
-    "branch": "feat/112-consolidate-master-template",
-    "observedRevision": "f4cd2dad89ff05d85d5887d46745dcc869b64e7b",
-    "activeIssue": 112,
-    "activeSpec": "context/specs/LV-201-consolidate-master-template.md",
-    "objective": "Move the repository-owned maximal application into template/ and merge currently duplicated module-owned source into that one source before later generator simplification.",
+    "branch": "feat/113-simplify-configuration-core",
+    "observedRevision": "a8593b3",
+    "activeIssue": 113,
+    "activeSpec": "context/specs/LV-202-simplify-configuration-core.md",
+    "objective": "Move useful preset defaults into core and remove the separate recipes package while preserving one shared configuration path.",
     "readFirst": [
       "AGENTS.md",
       "context/README.md",
-      "context/specs/LV-201-consolidate-master-template.md",
-      "context/docs/template.md",
+      "context/specs/LV-202-simplify-configuration-core.md",
+      "context/docs/product.md",
+      "context/docs/configuration.md",
       "context/docs/architecture.md",
-      "context/docs/repository-transition.md",
+      ".agents/contracts/product.yaml",
       ".agents/contracts/architecture.yaml",
-      ".agents/contracts/transition.yaml"
+      "packages/schema/src/recipe.ts",
+      "packages/core/src/recipe.ts"
     ],
     "doNot": [
       "reintroduce DigitalHerencia/Vibes",
-      "redesign the template internals for aesthetics",
-      "redesign the web UI in LV-201",
-      "rewrite the CLI in LV-201",
+      "add imaginary capabilities",
+      "expose fixed architecture as configuration",
+      "redesign the master template",
       "add new tests or validation systems",
       "claim checks that were not run"
     ]
