+++ .agents/execution/progress.json
@@ -3,10 +3,10 @@
   "status": "in-progress",
   "observedRevision": "f4cd2dad89ff05d85d5887d46745dcc869b64e7b",
   "observedDate": "2026-08-09",
-  "currentOutcome": "LV-201 relocates the sole maximal application source to template/ while retaining temporary module ownership metadata for compatibility.",
-  "activeIssue": 112,
-  "activeSpec": "context/specs/LV-201-consolidate-master-template.md",
-  "openIssuesObserved": 7,
+  "currentOutcome": "LV-202 removes the recipes package while preserving shared preset defaults in the generator core.",
+  "activeIssue": 113,
+  "activeSpec": "context/specs/LV-202-simplify-configuration-core.md",
+  "openIssuesObserved": 6,
   "openPullRequestsObserved": 0,
   "completedPriorRoadmap": "LV-101 through LV-110",
   "nextRoadmap": [
@@ -19,7 +19,7 @@
     "LV-207"
   ],
   "knownTransitionFacts": [
-    "packages/recipes still exists",
+    "preset defaults are owned by packages/core",
     "template is the sole complete application source",
     "templates/modules contains compatibility metadata only",
     "package.json publishes template plus temporary module metadata",
@@ -35,5 +35,5 @@
     "package smoke checks",
     "deployments"
   ],
-  "nextAction": "Finish LV-201 verification, open its focused pull request, and merge before starting LV-202."
+  "nextAction": "Finish LV-202 verification, open its focused pull request, and merge before starting LV-203."
 }
