+++ packages/core/src/index.ts
@@ -46,7 +46,7 @@ export {
   type ResolvedBuildSummary,
   type ResolvedRecipe,
 } from './recipe.js';
-export { getProductPreset, productPresets } from '@loaded-vibes/recipes';
+export { getProductPreset, productPresets } from './presets.js';
 export type {
   CapabilityId,
   Design,
