--- packages/recipes/package.json
@@ -1,12 +1,0 @@
-{
-  "name": "@loaded-vibes/recipes",
-  "version": "0.1.0",
-  "private": true,
-  "type": "module",
-  "exports": {
-    ".": "./src/index.ts"
-  },
-  "dependencies": {
-    "@loaded-vibes/schema": "workspace:*"
-  }
-}
