--- packages/recipes/src/index.ts
@@ -1,5 +1,0 @@
-export {
-  getProductPreset,
-  productPresets,
-  type ProductPreset,
-} from './presets.js';
